Create Decision Tree using IRIS dataset
The following example uses the iris data set. This dataset contains 3 classes of 150 instances each, where each class refers to the type of the iris plant. 
Our objective here is to find a tree, which can tell us if an Iris flower species belongs to one of following classes: setosa, versicolor or virginica.